+++ graphql-maven-plugin-logic/src/main/resources/templates/server_GraphQLDataFetchersDelegate.vm.java
@@ -6,10 +6,14 @@ import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
+import org.dataloader.BatchLoaderEnvironment;
 import org.dataloader.DataLoader;
 import org.reactivestreams.Publisher;
 
 import graphql.schema.DataFetchingEnvironment;
+#if($configuration.generateBatchLoaderEnvironment)
+import org.dataloader.BatchLoaderEnvironment;
+#end
 
 #foreach($import in $imports)
 import $import;
@@ -90,7 +94,7 @@ public interface ${dataFetcherDelegate.pascalCaseName} {
 	 *            A list of ${batchLoader.type.identifier.type.name}'s id
 	 * @return A list of ${batchLoader.type.identifier.type.name}s
 	 */
-	public List<${batchLoader.type.classSimpleName}> batchLoader(List<${batchLoader.type.identifier.type.classSimpleName}> keys);
+	public List<${batchLoader.type.classSimpleName}> batchLoader(List<${batchLoader.type.identifier.type.classSimpleName}> keys#if($configuration.generateBatchLoaderEnvironment), BatchLoaderEnvironment environment#end);
 
 #end
 }