+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateAllFieldCasesWithIdSubtypeImpl.java
@@ -8,6 +8,7 @@ import javax.annotation.Resource;
 
 import org.allGraphQLCases.server.AllFieldCasesWithIdSubtype;
 import org.allGraphQLCases.server.util.DataFetchersDelegateAllFieldCasesWithIdSubtype;
+import org.dataloader.BatchLoaderEnvironment;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -18,12 +19,19 @@ public class DataFetchersDelegateAllFieldCasesWithIdSubtypeImpl
 	DataGenerator generator;
 
 	@Override
-	public List<AllFieldCasesWithIdSubtype> batchLoader(List<UUID> keys) {
+	public List<AllFieldCasesWithIdSubtype> batchLoader(List<UUID> keys, BatchLoaderEnvironment environment) {
 		List<AllFieldCasesWithIdSubtype> list = new ArrayList<>(keys.size());
 
 		for (UUID id : keys) {
 			AllFieldCasesWithIdSubtype item = generator.generateInstance(AllFieldCasesWithIdSubtype.class);
 			item.setId(id);
+
+			// Let's manage the uppercase parameter, that was associated with this key
+			Boolean uppercase = (Boolean) environment.getKeyContexts().get(id);
+			if (uppercase != null && uppercase) {
+				item.setName(item.getName().toUpperCase());
+			}
+
 			list.add(item);
 		}