+++ graphql-maven-plugin/src/main/java/com/graphql_java_generator/mavenplugin/AbstractGenerateServerCodeMojo.java
@@ -4,8 +4,10 @@
 package com.graphql_java_generator.mavenplugin;
 
 import java.io.File;
+import java.util.Map;
 
 import org.apache.maven.plugins.annotations.Parameter;
+import org.dataloader.BatchLoaderEnvironment;
 
 import com.graphql_java_generator.plugin.conf.GenerateServerCodeConfiguration;
 import com.graphql_java_generator.plugin.conf.GraphQLConfiguration;
@@ -25,7 +27,62 @@ public abstract class AbstractGenerateServerCodeMojo extends AbstractGenerateCod
 		implements GenerateServerCodeConfiguration {
 
 	/**
-	 * Indicates whether the plugin should generate the JPA annotations, for generated objects, when in server mode.
+	 * <P>
+	 * (only for server mode) Indicates if the plugin should generate add the {@link BatchLoaderEnvironment} parameter
+	 * to the <I>batchLoader</I> methods, in DataFetchersDelegate. This parameter allows to get the context of the Batch
+	 * Loader, including the context associated to the id, when using the id has been added by the
+	 * {@link org.dataloader.DataLoader#load(Object, Object)} method.
+	 * </P>
+	 * <P>
+	 * For instance, if you have the method below, for a field named <I>oneWithIdSubType</I> in a DataFetcherDelegate:
+	 * </P>
+	 * 
+	 * <PRE>
+	 * &#64;Override
+	 * public CompletableFuture<AllFieldCasesWithIdSubtype> oneWithIdSubType(
+	 * 		DataFetchingEnvironment dataFetchingEnvironment, DataLoader<UUID, AllFieldCasesWithIdSubtype> dataLoader,
+	 * 		AllFieldCases source, Boolean uppercase) {
+	 * 	return dataLoader.load(UUID.randomUUID());
+	 * }
+	 * </PRE>
+	 * <P>
+	 * then, in the <I>AllFieldCasesWithIdSubtype</I> DataFetcherDelegate, you can retrieve the uppercase this way:
+	 * </P>
+	 * 
+	 * <PRE>
+	 * &#64;Override
+	 * public List<AllFieldCasesWithIdSubtype> batchLoader(List<UUID> keys, BatchLoaderEnvironment environment) {
+	 * 	List<AllFieldCasesWithIdSubtype> list = new ArrayList<>(keys.size());
+	 * 	for (UUID id : keys) {
+	 * 		// Let's manage the uppercase parameter, that was associated with this key
+	 * 		Boolean uppercase = (Boolean) environment.getKeyContexts().get(id);
+	 * 		if (uppercase != null && uppercase) {
+	 * 			item.setName(item.getName().toUpperCase());
+	 * 		}
+	 * 
+	 * 		// Do something with the id and the uppercase value
+	 * 	}
+	 * 	return list;
+	 * }
+	 * </PRE>
+	 * <P>
+	 * For more complex cases, you can store a {@link Map} with all the needed values, instead of just the parameter
+	 * value.
+	 * </P>
+	 * <P>
+	 * <B><I>Default value is false</I></B>
+	 * </P>
+	 */
+	@Parameter(property = "com.graphql_java_generator.mavenplugin.generateBatchLoaderEnvironment", defaultValue = GraphQLConfiguration.DEFAULT_GENERATE_BATCH_LOADER_ENVIRONMENT)
+	public boolean generateBatchLoaderEnvironment;
+
+	/**
+	 * <P>
+	 * (only for server mode) Indicates whether the plugin should generate the JPA annotations, for generated objects.
+	 * </P>
+	 * <P>
+	 * <B><I>Default value is false</I></B>
+	 * </P>
 	 */
 	@Parameter(property = "com.graphql_java_generator.mavenplugin.generateJPAAnnotation", defaultValue = GraphQLConfiguration.DEFAULT_GENERATE_JPA_ANNOTATION)
 	boolean generateJPAAnnotation;
@@ -111,6 +168,11 @@ public abstract class AbstractGenerateServerCodeMojo extends AbstractGenerateCod
 	public File getSchemaPersonalizationFile() {
 		return (GraphQLConfiguration.DEFAULT_SCHEMA_PERSONALIZATION_FILE.equals(schemaPersonalizationFile)) ? null
 				: new File(project.getBasedir(), schemaPersonalizationFile);
+	}
+
+	@Override
+	public boolean isGenerateBatchLoaderEnvironment() {
+		return generateBatchLoaderEnvironment;
 	}
 
 	@Override